+++ libs/testnorge-auto-config-dependency-analysis/src/main/java/no/nav/registre/testnorge/libs/autoconfigdependencyanalysis/config/AvhengighetsanalyseServiceClientCredential.java
@@ -1,4 +1,4 @@
-package no.nav.registre.testnorge.libs.autodependencyanalysis.config;
+package no.nav.registre.testnorge.libs.autoconfigdependencyanalysis.config;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Configuration;
