+++ libs/testnorge-auto-config-dependency-analysis/src/main/java/no/nav/registre/testnorge/libs/autoconfigdependencyanalysis/config/DependencyAnalysisAutoConfiguration.java
@@ -1,18 +1,29 @@
-package no.nav.registre.testnorge.libs.autodependencyanalysis.service;
+package no.nav.registre.testnorge.libs.autoconfigdependencyanalysis.config;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Service;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
+import org.springframework.context.annotation.Profile;
 
-import no.nav.registre.testnorge.libs.autodependencyanalysis.consumer.AvhengighetsanalyseServiceConsumer;
+import javax.annotation.PostConstruct;
+
+import no.nav.registre.testnorge.libs.autoconfigdependencyanalysis.consumer.AvhengighetsanalyseServiceConsumer;
+import no.nav.registre.testnorge.libs.oauth2.service.ClientCredentialGenerateWithoutLoginAccessTokenService;
 
 @Slf4j
-@Service
-public class ApplicationRegistrationService {
+@Configuration
+@Import({
+        AvhengighetsanalyseServiceClientCredential.class,
+        AvhengighetsanalyseServiceConsumer.class,
+        ClientCredentialGenerateWithoutLoginAccessTokenService.class
+})
+@Profile("prod")
+public class DependencyAnalysisAutoConfiguration {
     private final AvhengighetsanalyseServiceConsumer avhengighetsanalyseServiceConsumer;
     private final String applicationName;
 
-    public ApplicationRegistrationService(
+    public DependencyAnalysisAutoConfiguration(
             AvhengighetsanalyseServiceConsumer avhengighetsanalyseServiceConsumer,
             @Value("${application.name:${spring.application.name:#{null}}}") String applicationName
     ) {
@@ -20,8 +31,8 @@ public class ApplicationRegistrationService {
         this.applicationName = applicationName;
     }
 
-
-    public void register() {
+    @PostConstruct
+    public void init() {
         try {
             if (applicationName != null) {
                 avhengighetsanalyseServiceConsumer.registerApplication(applicationName);
